Nit - refactor naming styles of components
