+++ packages/components/src/lib/primitives/FLRadioGroup/index.spec.tsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react';
 
-import FLRadioGroup from './FLRadioGroup';
+import FLRadioGroup from '.';
 
 describe('FLRadioGroup', () => {
   it('should render successfully', () => {
