+++ packages/components/src/lib/primitives/index.ts
@@ -1,2 +1,2 @@
-export * from './FLRadioGroup/FLRadioGroup'
-export * from './FLTooltip/FLTooltip'
+export * from './FLRadioGroup'
+export * from './FLTooltip'
